requirements.txt
#Flask=2.2.2
prometheus-client #==0.14.1
pyyaml  #==6.0